refactor : refactor to hexagonal structure
