--- src/main/java/com/joel/users/api/controllers/AuthenticationController.java
@@ -1,24 +1,0 @@
-package com.joel.users.api.controllers;
-
-import com.joel.users.domain.dtos.request.UserRequestDTO;
-import com.joel.users.domain.dtos.response.UserDTO;
-import com.joel.users.domain.usecases.CreateUserUseCase;
-import jakarta.validation.Valid;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.*;
-
-@RequiredArgsConstructor
-@RestController
-@RequestMapping("/api/auth")
-public class AuthenticationController {
-
-    private final CreateUserUseCase createUserUseCase;
-
-    @PostMapping("/signup")
-    @ResponseStatus(HttpStatus.CREATED)
-    public UserDTO registerUser(@RequestBody @Valid UserRequestDTO userRequestDTO) {
-        return createUserUseCase.createUser(userRequestDTO);
-    }
-
-}
