--- src/main/java/com/joel/users/api/exceptionHandler/ApiExceptionHandler.java
@@ -1,273 +1,0 @@
-package com.joel.users.api.exceptionHandler;
-
-
-import com.fasterxml.jackson.databind.JsonMappingException;
-import com.fasterxml.jackson.databind.exc.InvalidFormatException;
-import com.fasterxml.jackson.databind.exc.PropertyBindingException;
-import com.joel.users.domain.exceptions.BusinessException;
-import com.joel.users.domain.exceptions.EntityInUseException;
-import com.joel.users.domain.exceptions.EntityNotFoundException;
-import lombok.NonNull;
-import lombok.RequiredArgsConstructor;
-import org.apache.commons.lang3.exception.ExceptionUtils;
-import org.hibernate.exception.ConstraintViolationException;
-import org.springframework.beans.TypeMismatchException;
-import org.springframework.context.MessageSource;
-import org.springframework.context.i18n.LocaleContextHolder;
-import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.HttpStatusCode;
-import org.springframework.http.ResponseEntity;
-import org.springframework.http.converter.HttpMessageNotReadableException;
-import org.springframework.security.authentication.InternalAuthenticationServiceException;
-import org.springframework.validation.BindingResult;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.context.request.WebRequest;
-import org.springframework.web.method.annotation.MethodArgumentTypeMismatchException;
-import org.springframework.web.servlet.NoHandlerFoundException;
-import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
-
-import java.time.OffsetDateTime;
-import java.util.List;
-import java.util.stream.Collectors;
-
-
-@RequiredArgsConstructor
-@ControllerAdvice
-public class ApiExceptionHandler extends ResponseEntityExceptionHandler {
-
-    public static final String MSG_GENERIC_ERROR_END_USER = "An unexpected internal system error has occurred. " +
-            "Try again and if the problem persists, contact your system administrator.";
-    public static final String ENTITY_IN_USE = "Entity cannot be deleted because it is in use";
-    public static final String INVALID_USERNAME_OR_PASSWORD = "Invalid username or password";
-
-    private final MessageSource messageSource;
-
-    @ExceptionHandler(EntityNotFoundException.class)
-    public ResponseEntity<?> handleEntityNotFound(EntityNotFoundException ex, WebRequest webRequest) {
-
-        HttpStatus status = HttpStatus.NOT_FOUND;
-        ProblemType problemType = ProblemType.RESOURCE_NOT_FOUND;
-        String detail = ex.getMessage();
-
-        Problem problem = createProblemBuilder(status, problemType, detail)
-                .userMessage(detail)
-                .build();
-
-        return handleExceptionInternal(ex, problem, new HttpHeaders(), status, webRequest);
-    }
-
-    @ExceptionHandler(BusinessException.class)
-    public ResponseEntity<?> handleBusiness(BusinessException ex, WebRequest webRequest) {
-        HttpStatus status = HttpStatus.BAD_REQUEST;
-        ProblemType problemType = ProblemType.BUSINESS_ERROR;
-        String detail = ex.getMessage();
-
-        Problem problem = createProblemBuilder(status, problemType, detail)
-                .userMessage(detail)
-                .build();
-
-        return handleExceptionInternal(ex, problem, new HttpHeaders(), status, webRequest);
-    }
-
-    @ExceptionHandler(EntityInUseException.class)
-    public ResponseEntity<?> handleEntityInUse(EntityInUseException ex, WebRequest webRequest) {
-        HttpStatus status = HttpStatus.CONFLICT;
-        ProblemType problemType = ProblemType.ENTITY_IN_USE;
-        String detail = ex.getMessage();
-
-        Problem problem = createProblemBuilder(status, problemType, detail)
-                .userMessage(detail)
-                .build();
-
-        return handleExceptionInternal(ex, problem, new HttpHeaders(), status, webRequest);
-    }
-
-
-    @ExceptionHandler(InternalAuthenticationServiceException.class)
-    public ResponseEntity<?> internalAuthenticationService(InternalAuthenticationServiceException ex, WebRequest webRequest) {
-        HttpStatus status = HttpStatus.UNAUTHORIZED;
-        ProblemType problemType = ProblemType.INVALID_PARAMETER;
-        String detail = ex.getMessage();
-
-        Problem problem = createProblemBuilder(status, problemType, detail)
-                .userMessage(INVALID_USERNAME_OR_PASSWORD)
-                .build();
-
-        return handleExceptionInternal(ex, problem, new HttpHeaders(), status, webRequest);
-    }
-
-
-    @Override
-    protected ResponseEntity<Object> handleHttpMessageNotReadable(
-            @NonNull HttpMessageNotReadableException ex,
-            @NonNull HttpHeaders headers,
-            @NonNull HttpStatusCode status,
-            @NonNull WebRequest request) {
-        Throwable rootCause = ExceptionUtils.getRootCause(ex);
-
-        if (rootCause instanceof InvalidFormatException) {
-            return handleInvalidFormatException((InvalidFormatException) rootCause, headers, (HttpStatus) status, request);
-        } else if (rootCause instanceof PropertyBindingException) {
-            return handlePropertyBindingException((PropertyBindingException) rootCause, headers, (HttpStatus) status, request);
-        }
-
-        ProblemType problemType = ProblemType.INCOMPREHENSIBLE_MESSAGE;
-        String detail = "The request body is invalid. Check for syntax errors.";
-
-        Problem problem = createProblemBuilder(status, problemType, detail)
-                .userMessage(detail)
-                .build();
-
-        return handleExceptionInternal(ex, problem, headers, status, request);
-    }
-
-    @ExceptionHandler(DataIntegrityViolationException.class)
-    public ResponseEntity<?> handleDataIntegrityViolation(DataIntegrityViolationException ex, WebRequest webRequest) {
-        HttpStatus status = HttpStatus.CONFLICT;
-        ProblemType problemType = ProblemType.ENTITY_IN_USE;
-        String detail = ENTITY_IN_USE;
-
-        Problem problem = createProblemBuilder(status, problemType, detail)
-                .userMessage(detail)
-                .build();
-
-        return handleExceptionInternal(ex, problem, new HttpHeaders(), status, webRequest);
-    }
-
-    @ExceptionHandler(ConstraintViolationException.class)
-    public ResponseEntity<?> handleDataIntegrityViolation(ConstraintViolationException ex, WebRequest webRequest) {
-        HttpStatus status = HttpStatus.BAD_REQUEST;
-        ProblemType problemType = ProblemType.INVALID_DATA;
-        String detail = ex.getMessage();
-
-        Problem problem = createProblemBuilder(status, problemType, detail)
-                .userMessage(detail)
-                .build();
-
-        return handleExceptionInternal(ex, problem, new HttpHeaders(), status, webRequest);
-    }
-
-    @Override
-    protected ResponseEntity<Object> handleTypeMismatch(
-            @NonNull TypeMismatchException ex, @NonNull HttpHeaders headers,
-            @NonNull HttpStatusCode status, @NonNull WebRequest request) {
-
-        if (ex instanceof MethodArgumentTypeMismatchException) {
-            return handleMethodArgumentTypeMismatch(
-                    (MethodArgumentTypeMismatchException) ex, headers, (HttpStatus) status, request);
-        }
-
-        return super.handleTypeMismatch(ex, headers, status, request);
-    }
-
-    @Override
-    protected ResponseEntity<Object> handleNoHandlerFoundException(
-            NoHandlerFoundException ex, @NonNull HttpHeaders headers,
-            @NonNull HttpStatusCode status, @NonNull WebRequest request) {
-
-        ProblemType problemType = ProblemType.RESOURCE_NOT_FOUND;
-        String detail = String.format("The resource %s, which you tried to access, is non-existent.", ex.getRequestURL());
-
-        Problem problem = createProblemBuilder(status, problemType, detail).build();
-
-        return handleExceptionInternal(ex, problem, headers, status, request);
-    }
-
-    @Override
-    protected ResponseEntity<Object> handleMethodArgumentNotValid(
-            MethodArgumentNotValidException ex, @NonNull HttpHeaders headers,
-            @NonNull HttpStatusCode status, @NonNull WebRequest request) {
-
-        ProblemType problemType = ProblemType.INVALID_DATA;
-        String detail = "One or more fields are invalid. Fill in correctly and try again.";
-
-        BindingResult bindingResult = ex.getBindingResult();
-
-        List<Problem.Object> problemObjects = bindingResult.getFieldErrors().stream()
-                .map(fieldError -> {
-
-                    String message = messageSource.getMessage(fieldError, LocaleContextHolder.getLocale());
-
-                    return Problem.Object.builder()
-                            .name(fieldError.getField())
-                            .userMessage(message)
-                            .build();
-                })
-                .collect(Collectors.toList());
-
-        Problem problem = createProblemBuilder(status, problemType, detail)
-                .userMessage(detail)
-                .objects(problemObjects)
-                .build();
-
-        return handleExceptionInternal(ex, problem, headers, status, request);
-    }
-
-    private ResponseEntity<Object> handleMethodArgumentTypeMismatch(
-            MethodArgumentTypeMismatchException ex, HttpHeaders headers,
-            HttpStatus status, WebRequest request) {
-
-        ProblemType problemType = ProblemType.INVALID_PARAMETER;
-
-        assert ex.getRequiredType() != null;
-        String detail = String.format("The URL parameter '%s' was assigned the value '%s', " +
-                        "which is of an invalid type. Please correct and enter a value compatible with the type %s.",
-                ex.getName(), ex.getValue(), ex.getRequiredType().getSimpleName());
-
-        Problem problem = createProblemBuilder(status, problemType, detail).build();
-
-        return handleExceptionInternal(ex, problem, headers, status, request);
-    }
-
-    private ResponseEntity<Object> handleInvalidFormatException(InvalidFormatException ex,
-                                                                HttpHeaders headers, HttpStatus status, WebRequest request) {
-
-        String path = joinPath(ex.getPath());
-
-        ProblemType problemType = ProblemType.INCOMPREHENSIBLE_MESSAGE;
-        String detail = String.format("The property '%s' received the value '%s'," +
-                        "which is of an invalid type. Correct and enter a value compatible with the type %s.",
-                path, ex.getValue(), ex.getTargetType().getSimpleName());
-
-        Problem problem = createProblemBuilder(status, problemType, detail)
-                .userMessage(MSG_GENERIC_ERROR_END_USER)
-                .build();
-        return handleExceptionInternal(ex, problem, headers, status, request);
-    }
-
-    private ResponseEntity<Object> handlePropertyBindingException(PropertyBindingException ex,
-                                                                  HttpHeaders headers, HttpStatus status, WebRequest request) {
-
-        String path = joinPath(ex.getPath());
-
-        ProblemType problemType = ProblemType.INCOMPREHENSIBLE_MESSAGE;
-        String detail = String.format("The property '%s' does not exist. " +
-                "Please correct or remove this property and try again.", path);
-
-        Problem problem = createProblemBuilder(status, problemType, detail)
-                .userMessage(MSG_GENERIC_ERROR_END_USER)
-                .build();
-
-        return handleExceptionInternal(ex, problem, headers, status, request);
-    }
-
-    private String joinPath(List<JsonMappingException.Reference> references) {
-        return references.stream()
-                .map(JsonMappingException.Reference::getFieldName)
-                .collect(Collectors.joining("."));
-    }
-
-    private Problem.ProblemBuilder createProblemBuilder(HttpStatusCode status,
-                                                        ProblemType problemType, String detail) {
-        return Problem.builder()
-                .timestamp(OffsetDateTime.now())
-                .status(status.value())
-                .type(problemType.getUri())
-                .title(problemType.getTitle())
-                .detail(detail);
-    }
-}
