--- src/main/java/com/joel/users/api/exceptionHandler/Problem.java
@@ -1,32 +1,0 @@
-package com.joel.users.api.exceptionHandler;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import com.fasterxml.jackson.annotation.JsonInclude;
-import lombok.Builder;
-import lombok.Getter;
-
-import java.time.OffsetDateTime;
-import java.util.List;
-
-@JsonInclude(JsonInclude.Include.NON_NULL)
-@Getter
-@Builder
-public class Problem {
-
-    private Integer status;
-    private String type;
-    private String title;
-    private String detail;
-    private String userMessage;
-    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "dd-MM-yyyy HH:mm:ss")
-    private OffsetDateTime timestamp;
-    private List<Object> objects;
-
-    @Getter
-    @Builder
-    public static class Object {
-        private String name;
-        private String userMessage;
-    }
-
-}
