+++ src/main/java/com/joel/users/application/dtos/response/UserDTO.java
@@ -1,0 +1,33 @@
+package com.joel.users.application.dtos.response;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+import com.fasterxml.jackson.annotation.JsonInclude;
+import lombok.*;
+
+import java.time.OffsetDateTime;
+import java.util.UUID;
+
+@Builder(toBuilder = true)
+@Getter
+@Setter
+@JsonInclude(JsonInclude.Include.NON_NULL)
+@AllArgsConstructor
+@NoArgsConstructor
+public class UserDTO {
+
+    private UUID userId;
+    private String username;
+    private String email;
+    private String fullName;
+    private String userStatus;
+    private String userType;
+    private String phoneNumber;
+    private String cpf;
+    private String imageUrl;
+
+    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "dd-MM-YYYY HH:mm:ss")
+    private OffsetDateTime creationDate;
+
+    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "dd-MM-YYYY HH:mm:ss")
+    private OffsetDateTime updateDate;
+}
