+++ src/main/java/com/joel/users/application/ports/ShowUserUseCase.java
@@ -1,0 +1,10 @@
+package com.joel.users.application.ports;
+
+import com.joel.users.application.dtos.response.UserDTO;
+
+import java.util.UUID;
+
+public interface ShowUserUseCase {
+
+    UserDTO findById(UUID userId);
+}
