--- src/main/java/com/joel/users/domain/dtos/request/UserRequestDTO.java
@@ -1,39 +1,0 @@
-package com.joel.users.domain.dtos.request;
-
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Size;
-import lombok.*;
-import org.hibernate.validator.constraints.br.CPF;
-
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter
-@Setter
-public class UserRequestDTO {
-
-    @NotBlank
-    @Size(min = 6, max = 50)
-    private String username;
-
-    @NotBlank
-    @Email
-    private String email;
-
-    @NotBlank
-    @Size(min = 8, max = 100)
-    private String password;
-
-    @NotBlank
-    @Size(min = 10, max = 150)
-    private String fullName;
-
-    @NotBlank
-    @Size(min = 10, max = 15)
-    private String phoneNumber;
-
-    @CPF
-    @NotBlank
-    private String cpf;
-}
