--- src/main/java/com/joel/users/domain/dtos/response/UserDTO.java
@@ -1,31 +1,0 @@
-package com.joel.users.domain.dtos.response;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import lombok.*;
-
-import java.time.OffsetDateTime;
-import java.util.UUID;
-
-@Builder(toBuilder = true)
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-public class UserDTO {
-
-    private UUID userId;
-    private String username;
-    private String email;
-    private String fullName;
-    private String userStatus;
-    private String userType;
-    private String phoneNumber;
-    private String cpf;
-    private String imageUrl;
-
-    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "dd-MM-YYYY HH:mm:ss")
-    private OffsetDateTime creationDate;
-
-    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "dd-MM-YYYY HH:mm:ss")
-    private OffsetDateTime updateDate;
-}
