+++ src/main/java/com/joel/users/domain/repositories/UserRepository.java
@@ -1,11 +1,17 @@
 package com.joel.users.domain.repositories;
 
 import com.joel.users.domain.entities.User;
-import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.util.Optional;
 import java.util.UUID;
 
-public interface UserRepository extends JpaRepository<User, UUID> {
+public interface UserRepository {
+
+    Optional<User> findById(UUID id);
+
+    User save(User user);
+
+    void deleteById(UUID id);
 
     boolean existsByUsername(String username);
