--- src/main/java/com/joel/users/domain/usecases/CreateUserUseCase.java
@@ -1,9 +1,0 @@
-package com.joel.users.domain.usecases;
-
-import com.joel.users.domain.dtos.request.UserRequestDTO;
-import com.joel.users.domain.dtos.response.UserDTO;
-
-public interface CreateUserUseCase {
-
-    UserDTO createUser(UserRequestDTO userRequestDTO);
-}
