--- src/main/java/com/joel/users/domain/usecases/ShowUserUseCase.java
@@ -1,10 +1,0 @@
-package com.joel.users.domain.usecases;
-
-import com.joel.users.domain.dtos.response.UserDTO;
-
-import java.util.UUID;
-
-public interface ShowUserUseCase {
-
-    UserDTO findById(UUID userId);
-}
