+++ src/main/java/com/joel/users/infrastructure/adapters/jpa/UserJpaRepository.java
@@ -1,0 +1,14 @@
+package com.joel.users.infrastructure.adapters.jpa;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.UUID;
+
+public interface UserJpaRepository extends JpaRepository<UserEntity, UUID> {
+
+    boolean existsByUsername(String username);
+
+    boolean existsByEmail(String email);
+
+    boolean existsByCpf(String cpf);
+}
